在main中直接初始化republisher和物料的mesh节点
